refactor: organize things for html-renderer
