+++ packages/react-renderer/src/RichText.tsx
@@ -1,7 +1,5 @@
 import React, { Fragment } from 'react';
 import {
-  RichTextProps,
-  NodeRendererType,
   ElementNode,
   EmptyElementsToRemove,
   Node,
@@ -9,12 +7,15 @@ import {
   isText,
   isEmpty,
   elementTypeKeys,
-  EmbedReferences,
-  getArrayOfElements,
 } from '@graphcms/rich-text-types';
 
 import { defaultElements } from './defaultElements';
 import { RenderText } from './RenderText';
+import { RichTextProps } from './types';
+
+function getArrayOfElements(content: RichTextProps['content']) {
+  return Array.isArray(content) ? content : content.children;
+}
 
 function RenderNode({
   node,
@@ -24,8 +25,8 @@ function RenderNode({
 }: {
   node: Node;
   parent: Node | null;
-  renderers?: NodeRendererType;
-  references?: EmbedReferences;
+  renderers?: RichTextProps['renderers'];
+  references?: RichTextProps['references'];
 }) {
   if (isText(node)) {
     let text = node.text;
@@ -69,8 +70,8 @@ function RenderElement({
   references,
 }: {
   element: ElementNode;
-  renderers?: NodeRendererType;
-  references?: EmbedReferences;
+  renderers?: RichTextProps['renderers'];
+  references?: RichTextProps['references'];
 }) {
   const { children, type, ...rest } = element;
   const { nodeId, nodeType } = rest;
@@ -173,7 +174,7 @@ function RenderElement({
 
   const elementNodeRenderer = isEmbed
     ? elementToRender
-    : renderers?.[elementTypeKeys[type] as keyof NodeRendererType];
+    : renderers?.[elementTypeKeys[type] as keyof RichTextProps['renderers']];
 
   const NodeRenderer = elementNodeRenderer as React.ElementType;
 
@@ -233,11 +234,11 @@ export function RichText({
   // Asset object will be completly overriden by the resolvers. We need to keep
   // the default elements for the Asset that hasn't been writen.
   const assetRenderers = {
-    ...defaultElements.Asset,
+    ...defaultElements?.Asset,
     ...resolvers?.Asset,
   };
 
-  const renderers: NodeRendererType = {
+  const renderers: RichTextProps['renderers'] = {
     ...defaultElements,
     ...resolvers,
     Asset: assetRenderers,
@@ -257,14 +258,14 @@ export function RichText({
     return <Fragment />;
   }
 
-  const elements = getArrayOfElements(content);
-
   /*
     Checks if there's a embed type inside the content and if the `references` prop is defined
 
     If it isn't defined and there's embed elements, it will show a warning
-  */
+    */
   if (__DEV__) {
+    const elements = getArrayOfElements(content);
+
     const embedElements = elements.filter(element => element.type === 'embed');
 
     if (embedElements.length > 0 && !references) {
@@ -276,7 +277,7 @@ export function RichText({
 
   return (
     <RenderElements
-      content={elements}
+      content={content}
       renderers={renderers}
       references={references}
     />
