--- packages/types/src/util/getArrayOfElements.ts
@@ -1,5 +1,0 @@
-import { RichTextProps } from '..';
-
-export function getArrayOfElements(content: RichTextProps['content']) {
-  return Array.isArray(content) ? content : content.children;
-}
